requirements.txt
pynacl==1.5.0
colorama==0.4.6
